Haciendo controladores

+++ src/main/java/eus/fpsanturtzilh/repositories/MaterialakRepository.java
@@ -1,13 +1,11 @@
 package eus.fpsanturtzilh.repositories;
 
 import eus.fpsanturtzilh.models.Materialak;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
 public interface MaterialakRepository extends JpaRepository<Materialak, Integer> {
  
-	Optional<Materialak> findById(Long id);
+    Optional<Materialak> findById(Integer id);
 }
-
